Move subtask update logic that was only needed for tests into the tests that needed it.
